+++ src/images/decoder.h
@@ -61,10 +61,10 @@ public:
 	PixelFormat getFormat() const;
 
 	/** Return the number of mip maps contained in the image. */
-	uint32 getMipMapCount() const;
+	size_t getMipMapCount() const;
 
 	/** Return a mip map. */
-	const MipMap &getMipMap(uint32 mipMap) const;
+	const MipMap &getMipMap(size_t mipMap) const;
 
 	/** Return TXI data, if embedded in the image. */
 	virtual Common::SeekableReadStream *getTXI() const;
